refactor(transfers): own terminal finalization
